REFACTOR: mock API 컴포넌트 infrastructure로 이동(#52)
