+++ src/main/java/kr/hhplus/be/server/dataPlatform/infrastructure/persistence/DataPlatformClient.java
@@ -1,16 +1,17 @@
-package kr.hhplus.be.server.application.dataPlatform;
+package kr.hhplus.be.server.dataPlatform.infrastructure.persistence;
 
-import kr.hhplus.be.server.domain.payment.event.PaymentSuccessEvent;
+import kr.hhplus.be.server.payment.domain.event.PaymentSuccessEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-@Service
+@Component
 @RequiredArgsConstructor
 @Slf4j
-public class DataPlatformService {
+public class DataPlatformClient {
 
     public void sendPaymentInfo(PaymentSuccessEvent event) {
         log.info("데이터플랫폼 Mock API 호출: {}", event.getPaymentId());
     }
+
 }
