+++ src/main/java/kr/hhplus/be/server/payment/application/event/PaymentEventListener.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.payment.application.event;
 
-import kr.hhplus.be.server.application.dataPlatform.DataPlatformService;
+import kr.hhplus.be.server.dataPlatform.infrastructure.persistence.DataPlatformClient;
 import kr.hhplus.be.server.payment.domain.event.PaymentSuccessEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -11,10 +11,10 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @RequiredArgsConstructor
 public class PaymentEventListener {
 
-    private final DataPlatformService dataPlatformService; // 데이터플랫폼 연동용 서비스
+    private final DataPlatformClient dataPlatformClient;
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void handlePaymentSuccessEvent(PaymentSuccessEvent event) {
-        dataPlatformService.sendPaymentInfo(event);
+        dataPlatformClient.sendPaymentInfo(event);
     }
 }
